Changed processValue_t to int32_t, so that ONOFF works, but PID may be broken now

+++ include/Types.h
@@ -10,7 +10,7 @@
 
 #include <stdint.h>
 
-typedef uint16_t processValue_t;
+typedef int32_t processValue_t;
 inline processValue_t fivePercent(processValue_t in){return in / 20;}
 inline processValue_t ninetyFivePercent(processValue_t in){return (in - in / 20);}
 #define LIB_CONTROL_MINUS_LN05 3
